+++ functions/Search.fnc.php
@@ -50,9 +50,18 @@ function Search( $type, $extra = null )
 					if ( ! empty( $_REQUEST['school_id'] )
 						&& $_REQUEST['school_id'] != UserSchool() )
 					{
-						$_SESSION['UserSchool'] = DBGetOne( "SELECT ID FROM schools
-							WHERE SYEAR='" . UserSyear() . "'
-							AND ID='" . (int) $_REQUEST['school_id'] . "'" );
+						$user_schools = trim( (string) User( 'SCHOOLS' ), ',' ) ?
+							explode( ',', trim( User( 'SCHOOLS' ), ',' ) ) : '';
+
+						// Check user is in student's school.
+						if ( ! $user_schools
+							|| in_array( $_REQUEST['SCHOOL_ID'], $user_schools ) )
+						{
+							// Fix HACKING ATTEMPT in SetUserStudentID() when "Search All Schools" checked
+							$_SESSION['UserSchool'] = DBGetOne( "SELECT ID FROM schools
+								WHERE SYEAR='" . UserSyear() . "'
+								AND ID='" . (int) $_REQUEST['school_id'] . "'" );
+						}
 					}
 
 					SetUserStudentID( $_REQUEST['student_id'] );